# util/deephash: use unsafe.Pointer instead of reflect.Value